MC normalisation and global variable for configuration class.
MC normalisation: integrated luminosity is read from the data catalog files and the MC samples are normalised to this value. A parameter mcYieldFactorScale can be used to apply an extra scale factor in case only a fraction of the data sample has been processed following to the use of the maxEvents option.

Configuration: cfg is now global, so the same instance can be used from other places. Using the same instance prevent multiple printout of the configuration.