Changed fake correction method with proper statistical uncertainty calculation
Fake were estimated from MC scaled by ratio of integral of data and simulation histograms. The code was including a mistake in the uncertainty calculation. Now a fake rate is estimated from MC for each bin of the data histogram, which is used to correct the data (scaled by (1-fake)).