Fix Hugo's bug
Symptoms: pruner would crash when declaring pruners in files with
a name < "pruner.o" in alphabetical order. The root cause is a
combination of undefined behaviour (static initialization order)
and object file ordering in the Makefile (from the *.o glob).

Fix it by converting daughters_ to a pointer and performing the
initialization explicitely. A function-scope static variable is
used as an intermediate step in order to prevent further undefined
 behaviour.