Update of the 2016preAPV Catalogs
The xsec of each process has been updated with the latest values on POG Twikis or XSDB.
The xsec can now be found in xsec_2016preAPV.txt (just to keep track, if you want to change a xsec in the code, you have to do it in the corresponding Catalog).
The configuration file is up to date with the new Catalogs.
Some Catalogs have been renamed for standardization.